requirements.txt
numpy==1.13.3
pillow==6.2.0
Keras==2.1.5
tensorflow-gpu==1.7.0
tensorflow-tensorboard==0.4.0rc3
h5py==2.7.1
spacy==2.0.3
nltk==3.4.5
scikit-learn==0.19.1
sacred==0.7.2
pandas==0.22.0
